% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLselex.R
\name{allcatch}
\alias{allcatch}
\title{allcatch()}
\usage{
allcatch(stock)
}
\arguments{
\item{stock}{class FLStock}
}
\value{
FLStock
}
\description{
Function to assign all discards to landings for FLBRP refs
}
