% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLselex.R
\name{as.ogive}
\alias{as.ogive}
\title{as.ogive()}
\usage{
as.ogive(object)
}
\arguments{
\item{object}{selex FLpar or output from fitselex}

\item{nyears}{number of end years for reference}
}
\value{
FLStock
}
\description{
Function to set fbar range to F = max(Fa)
}
