% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLselex.R
\name{fabs}
\alias{fabs}
\title{Compute instantaneous F, such that  F_a = Fabs*Sa}
\usage{
fabs(stock, nyears = 3, year = NULL)
}
\arguments{
\item{stock}{Input FLStock object.}

\item{nys}{numbers of last years to compute selectivity}
}
\value{
value Fabs
}
\description{
Compute instantaneous F, such that  F_a = Fabs*Sa
}
