% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelexPlots.R
\name{Fselex}
\alias{Fselex}
\title{Fselex}
\usage{
Fselex(brps, what = c("Fref", "Fmsy", "F0.1"))
}
\arguments{
\item{brps}{output from brp.selex()}

\item{what}{type of F c("Fref","Fmsy","F0.1"), ref is by default Fcur}
}
\value{
data.frame(F,rel.yield,rel.ssb)
}
\description{
Returns Table with trade-offs between relative Catch and SSB as a function Selectivity
}
