% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLselex.R
\name{invVB}
\alias{invVB}
\title{invVB()}
\usage{
invVB(L, vBPar)
}
\arguments{
\item{L}{}
}
\value{
one value used to calculate the age using von Bertalanffy
}
\description{
Function to caculate the age corresponding to the age provided as input
}
\author{
Michael Gras
}
