% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLselexLen.R
\name{lopt}
\alias{lopt}
\title{lopt()}
\usage{
lopt(stock, nyears = 3)
}
\arguments{
\item{stock}{class FLStockLen}
}
\value{
FLQuant with annual spr0y
}
\description{
Function to compute lopt, the length where an unfished cohort attains maximum biomass
}
\author{
Michael Gras & Henning Winker
}
