% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelexPlots.R
\name{mleg}
\alias{mleg}
\title{mleg()
function to adjust ggplot legend for multiplots}
\usage{
mleg(size = 0.5, height = 0.5, width = 0.5, title = 6, text = 6)
}
\arguments{
\item{size}{legend.key.size (cm)}

\item{height}{legend.key.height (cm)}

\item{width}{legend.key.width (cm)}

\item{titel}{legend title size}

\item{test}{legend text size}
}
\value{
legformat
}
\description{
mleg()
function to adjust ggplot legend for multiplots
}
