% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLselex.R
\name{par2sa}
\alias{par2sa}
\title{par2sa()}
\usage{
par2sa(pars, object, nyears = 3)
}
\arguments{
\item{pars}{selexpars}

\item{object}{FLStock or FLQuant of Sa}

\item{nyears}{number of end years for reference}
}
\value{
FLStock
}
\description{
Function to convert selex pars to Sa (FLQuants)
}
