% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelexPlots.R
\name{ploteqselex_2}
\alias{ploteqselex_2}
\title{ploteqselex_2()
Isopleth plot of Selectivity vs F}
\usage{
ploteqselex(
  brps,
  fit,
  Fmax = 2,
  panels = NULL,
  ncol = NULL,
  colours = NULL,
  Ftrg = c("none", "msy", "f0.1")
)
}
\arguments{
\item{brps}{output from brp.selex()}

\item{fit}{output from }

\item{Fmax}{upper possible limit of  F range}

\item{panels}{choice of plots 1:4
\itemize{
  \item 1  F over Relative Yield
  \item 2  F over SRP or SSB (requires ssr)
  \item 3  Isopleth Yield
  \item 4  Isopleth SPR or SSB (requires ssr)
}}

\item{ncol}{number of columns}

\item{colours}{optional, e.g. terrain.col, rainbow, etc.}

\item{Ftrg}{option to dynamic Ftrg=c("none","fmsy","f0.1")
\itemize{
  \item "none"  
  \item "msy"  
  \item "f0.1"
}}
}
\value{
ggplot
}
\description{
ploteqselex()
Isopleth plot of Selectivity vs F
}
