% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelexPlots.R
\name{plotprjselex}
\alias{plotprjselex}
\title{plotprjselex()
Projection plot of FLStocks}
\usage{
plotprjselex(
  object,
  panels = NULL,
  ncol = NULL,
  colours = NULL,
  nyears = NULL,
  max.discrete = 10
)
}
\arguments{
\item{object}{FLStocks output from selex.forward/selex.backtest()}

\item{panels}{choice of plots 1:10
\itemize{
  \item 1  spawning biomass (SSB)
  \item 2  catch
  \item 3  Fjuv/F 
  \item 4  Percentage juveniles in catch in numbers
  \item 5  Harvest Rate = Catch/Vuln.Bio
  \item 6  vuln.bio 
  \item 7  fishing mortality F
  \item 8  total biomass
  \item 9  Frec/F (Vasilakopoulos et al. 2020)
  \item 10 Percentage in catch >= aopt
}}

\item{ncol}{number of columns}

\item{colours}{optional, e.g. terrain.col, rainbow, etc.}

\item{nyears}{number of years back from assessment year shown}

\item{max.discrete}{determines max selectivities after which plot legend becomes continous}
}
\value{
ggplot
}
\description{
plotprjselex()
Projection plot of FLStocks
}
