% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelexPlots.R
\name{plotselage}
\alias{plotselage}
\title{plots mean selectivity at age Sa across selected years}
\usage{
plotselage(stock, nyears = 5, year = NULL)
}
\arguments{
\item{stock}{Input FLStock object.}

\item{nyears}{numbers of last years to compute selectivity}

\item{year}{specific years (will overwrite nyears)}
}
\value{
FLQuant of Selectivity-at-age Sa
}
\description{
plots mean selectivity at age Sa across selected years
}
