% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelexPlots.R
\name{plotselex}
\alias{plotselex}
\title{plots mean selectivity at age Sa across selected years}
\usage{
plotselex(
  sel,
  Sa = NULL,
  obs = NULL,
  compounds = FALSE,
  colours = NULL,
  legend.title = "S50",
  max.discrete = 8
)
}
\arguments{
\item{sel}{selexpars FLPars(s) or Sa FLquants or output from fitselex()}

\item{Sa}{observed selectivity-at-age (FLQuant) or FLStock}

\item{obs}{show observations if TRUE}

\item{compounds}{option to show selex compounds}

\item{legend.title}{option for customized legend.title for discrete scenarios only}

\item{max.discrete}{determines max selectivities after which plot legend becomes continous}
}
\value{
FLQuant of Selectivity-at-age Sa
}
\description{
plots mean selectivity at age Sa across selected years
}
