% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLselex.R
\name{s50}
\alias{s50}
\title{approximates Age-at-50-selectivity from Sa}
\usage{
s50(Sa)
}
\arguments{
\item{Sa}{selectivity at age Sa = selage(stock)}
}
\value{
s50
}
\description{
approximates Age-at-50-selectivity from Sa
}
