% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLselex.R
\name{selage}
\alias{selage}
\title{coverts harvest() and catch.sel() selectivity at age Sa}
\usage{
selage(stock, nyears = 3, year = NULL, maturity = FALSE)
}
\arguments{
\item{stock}{Input FLStock object or FLQuant.}

\item{nyears}{numbers of last years to compute selectivity}

\item{year}{option to specify year range, overwrites nyears}

\item{maturity}{if TRUE selage() extract mat_at_age}
}
\value{
FLQuant of Selectivity-at-age Sa
}
\description{
coverts harvest() and catch.sel() selectivity at age Sa
}
