% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLselex.R
\name{selex}
\alias{selex}
\title{computes selex curves}
\usage{
selex(Sa, pars)
}
\arguments{
\item{Sa}{selectivity at age Sa = selage(stock)}

\item{pars}{selexpars S50, S95, Smax, Dcv, Dmin}
}
\value{
FLquants selex predictions
}
\description{
computes selex curves
}
