% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLselex.R
\name{varselex}
\alias{varselex}
\title{function to dynamically vary selex parameters}
\usage{
varselex(
  pars,
  stock,
  step = 0.1,
  amin = NULL,
  amax = NULL,
  nyears = 3,
  type = c("crank", "shift", "dynamic"),
  return = c("Pars", "Sa")
)
}
\arguments{
\item{pars}{5 selex parameters of class FLPar}

\item{stock}{optional stock object for tuning of age range}

\item{step}{step size of change in one or several pars}

\item{amin}{start of S50}

\item{amax}{end of S50, required if stock = NULL}

\item{nyears}{end years for computing aopt() as automated amax limit}

\item{type}{option of selectivity change "crank","shift" or "dynamic"}

\item{return}{type of returned object FLPars or FLQuants}
}
\value{
selex parameters (FLPars)
}
\description{
function to dynamically vary selex parameters
}
